html/semantics/forms/form-submission-0/form-submission-does-not-cancel-parsing-synchronously.html

<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = t.step_func_done(() => assert_equals(i.contentDocument.body.innerText, "PASS"));

  var i = document.createElement("iframe");
  i.id ="i";
  i.name = "i";
  i.src = "about:blank";
  document.body.appendChild(i);

  i.contentDocument.open();
  var form = document.body.appendChild(document.createElement("form"));
  form.action = "javascript:'PASS'";
  form.target = "i";
  form.submit();
  i.contentDocument.write("FAIL");
});
</script>
</body>
